work in progress
1. working on off opengl - rendering all triangles.
2. next to calculate normal, left-hand rule to determine if visible
3. next to add phong shading

+++ app/src/main/java/com/ctyeung/openglex/off/MyKnotsView.kt
@@ -89,7 +89,7 @@ class MyKnotsView(
                     return false;
                 }
                 // retrieve vertex
-                val vtx1 = _vertices[vIndex].let { PointF3D(it.x-center.x, it.y-center.y, it.z-center.z) }
+                val vtx1 = _vertices[vIndex.toInt()].let { PointF3D(it.x-center.x, it.y-center.y, it.z-center.z) }
 
                 val y = cos(radX) * vtx1.y - sin(radX) * vtx1.z
                 val z = sin(radX) * vtx1.y + cos(radX) * vtx1.z
